Avoid catching ErrnoException directly in ParcelFileDescripotorRewinder
ErrnoException requires API level 21. Caught exceptions are apparently
initialized at class load time, rather than at runtime. As a result, any
usage of the class will fail on older API levels. Guarding the specific
block of code is insufficient.

To preserve the external API and meet the OS requirement, I've moved the
logic that deals with Os.lseek into a static inner class that shouldn't
need to be initialized until an instance of ParcelFileDescriptorRewinder
is actually created. Since we guard creation of
ParcelFileDescriptorRewinder with #isSupported(), we should be safe.

PiperOrigin-RevId: 279987765