Add support for width / height to the SvgDecoder
The sample SvgDecoder does not currently use parameters with and height
that are passed into the decode() method. This can be misleading for
Glide users and causes basic functionality such as image scaling to not
work properly. This change adds code to change the SVG document width
and height according to the provided parameters.