Deduplicate HTTP Uri ModelLoaders
Glide had two essentially identical ModelLoaders: UrlUriLoader and
HttpUriLoader. They both were being used to defer HTTP/HTTPS Uris
to another ModelLoader.

Best-case scenario, this was a harmless distraction; but in some bad
cases, it was possible for both ModelLoaders to fire, duplicating work.

With this change, only one of the two HTTP Uri ModelLoaders are entered
into the registry. The one no longer being used has been deprecated as well.

I chose to deprecate HttpUriLoader because it can be implemented as a
UrlUriLoader (but not vice versa, due to generics). To demonstrate that
it still behaves the same, the test (which now targets UrlUriLoader) is
unchanged but still passes.