Reset mark position after skipping past the mark limit. Any later reset will throw instead of leaving the stream in a bad/undefined state.
Previously, this would cause us to lose track of the bytes that were skipped. For PNGs, ExifInterface skips past the entire data block, which for some files can exceed the buffer limit. We would return back to the buffer location, decode the image partially and then try to load more from the wrapped stream - which would lead to a quick EOF

This is similar to what fillbuf does when we push past the mark limit.

PiperOrigin-RevId: 349315836