Rollforward retry adding pending RequestManagerFragments.
This attempt includes a fix to handle starting requests before Activity onCreate during Activity re-creation.

When an Activity is re-created, the old request manager fragment is retained. Normally that's fine because when we start the request we look up the old fragment and use it instead of adding a new one. However if we manager to start the request before Activity.onCreate, the FragmentManager's Fragments will not be populated and we won't be able to find the previous Fragment. That leads us to create a new Fragment, associate the RequestManager with the new Fragment and add it in a posted Handler message. When that message runs, Activity.onCreate will have finished and we'll find both our new Fragment and the old Fragment.

To handle this case, I've narrowed the assertion to only throw if there are requests associated with the existing Fragment. If there are requests associated with both Fragments, then it's unsafe to remove both and we're in a weird undefined state. If there are no requests associated with the old/existing Fragment, then I just remove it and replace it with the new one.

PiperOrigin-RevId: 409253061