Start lifecycles not RequestManagers in RequestManagerRetriever.
This was originally added to fix a memory leak when we registered the connectivity receiver globally, but not the RequestManager. We've since started registering the RequestManager globally, so this fix no longer works. Neither fix should be necessary if we can make sure we use the correct Fragment here anyway, which I think we've done.

PiperOrigin-RevId: 410075941