Rollforward retry adding pending RequestManager Fragments.
This time I've made sure we look for pending fragments before we try to look up any previously added fragments. Prior to this change, we might add a pending fragment but then successfully look up a Fragment from an old instance of the Activity in a subsequent request if that request occurred after the old Fragment was re-added but before our pending Fragment was added. Now we will only ever use the pending Fragment if one has been added.

PiperOrigin-RevId: 414768848