Use android.compileSdkVersion instead of android.compileSdk
While you can set the sdk integer with compileSdk, you don't appear to
be able to read it from the project using that shortcut. Instead reads
still need to be done via the not yet deprecated compileSdkVersion.

Fixes a regression introduced in
https://github.com/bumptech/glide/pull/4828. I'm not clear on why the
presubmits did not catch the issue then, but started to after that cl.
I've explicitly started running the failing tasks in the hopes of
avoiding this in the future.