Add support for parsing LibraryGlideModules.
Every round now writes a new Index containing all LibraryGlideModules in that round. When we encounter an AppGlideModule, we read all previously read Index classes, extract the LibraryGlideModules they point to, and write a merged AppGlideModule that calls the developer's AppGlideModule and all of the LibraryGlideModules.

Excluding LibraryGlideModules is not supported yet. The library is still in a pre-release state.

Progress towards #4492

PiperOrigin-RevId: 461946515