Use release instead of debug variant, fixup benchmark package.
From Glide's perspective these two variants are exactly the same. We don't do any different configuration for release vs debug. However, the Kotlin binary-compatibility-validator only works on the release variant and it doesn't appear to be configurable to work on a different variant. To use that plugin, we need to have a release variant.

The net output here should be the same as before, just with Relase* task names instead of Debug* task names.

The benchmark project changes seem to be unrelated, but I think they're
actually required by some additional checks that the tools run by
default on release variants but not debug variants.