Add a Compose API for Glide.
This is relatively early stage at this point. The API is likely to change in
breaking ways. Ideally the migrations are not large, but I'd expect a
few iterations.

There's already a TODO to improve the Prealoader API and make the
GlideImage API less reliant on GlideBuilder.

Early feedback and usage would be welcome, with the caveat that it'll
mean some migration work down the road.

PiperOrigin-RevId: 475696354