Deprecate .theme()
We should now be able to pull the correct theme from the Context from
the associated RequestManager, which makes this method largely not
useful.

The only exception would be an attempt to do a themed loading using the
Application Context where a theme of some Fragment or Activity is
available. I assume that case is relatively rare and it's probably
better to just do the load with the Fragment or Activity Context anyway.