Enable DirectResourceLoader in Glide
The new loaders respect themes, dark mode, screen sizes and other resource attributes that are not supported by contentResolver.openInputStream. We can only use them reliably with resource ids and uris that are owned by the same package as the resource, so the legacy loaders will continue to live for other packages. There are also some edge cases, like videos in raw resources, that are not supported by the direct resource loaders.

This will result in a behavior change for applications. Resources that were previously loaded with default attributes will now be loaded instead with the appropriate theme / dark mode / size etc.

PiperOrigin-RevId: 509304116