Add integration tests for ksp library modules.
I've added a new module ksp:integrationtest so that we can keep the
pollution from the dependent module out of the normal library glide
module tests. There's more detail in the documentation for the
build.gradle file in ksp:integrationtests.

This does at least make sure that if a dependent module is added, it is
succesfully included by ksp. Previously the tests only tested either
library modules compiled with the app glide module or library modules
compiled by ksp. This new set of tests now tests library modules
compiled by java. If we ever compile a module with ksp, we should add
that module to these tests as well so that we have integration tests for
both types of modules. For now all modules are compiled only with the
java processor.

In an ideal world we'd be able to just run the Java annotation processor
using kotlin compile testing the same way we are with ksp. The library
seems to support it. However I can't figure out a way to get the javax
annotations and kotlin and android all available in a single gradle
module. For now using a separate module is reasonable (if somewhat
clunky) compromise.