Update Glide to 5.0, java 1.8, add a Kotlin based decode job
It's been a long time since we've done a major version upgrade. Some of
the most incomprehensible code in Glide is in DecodeJob, where we
minimize object allocations with object pooling and confusingly
recursive Generator like objects. Now that dalvik is pretty much
entirely a thing of the past, object allocations seem to be a very small
or even entirely non-existent performance problem.

To fix this I've added a more or less direct re-implementation of
DecodeJob in Kotlin. Except that instead of generators, it relies on
Kotlin's async await. Even as an otherwise minor re-write, it seems like
a significant win for readability and maintainability.

While it does seem to be theoretically possible to use Kotlin and
continue to target Java 7, in practice it's hard to do. Many of the
dependencies Glide uses have updated to require java 8. Avoiding those
upgrades would mean using older versions of these dependencies. It's
also simply been an awfully long time since the Android ecosystem
migrated to Java 7. Glide's codebase and several existing dependencies
are stuck at old versions. It seems past time to move to a more modern
language version.