Run regression tests when resources are present or we're regenerating.
Previously the regression tests would run always, which means that
they'd fail 100% of the time for any device, emulator or SDK combination
that we hadn't previously generated resources for.

By ignoring these tests when the resources are missing and we aren't
trying to generate them, we make the emulator test output much more
useful for additional device types or emulators (ie those used for
testing features, but not in our regular test suite).