Register a BitmapDrawableEncoder in ErrorHandlingTest
Right now this test relies on the fact that we happen to attempt to decode Bitmaps before BitmapDrawables. If we were instead to decode BitmapDrawables first, the test would fail because the FailEncoder we register would never be called.

To make the test more robust to order changes, I've registered an encoder for both BitmapDrawable and Bitmap, which are the two types we could plausibly decode here.

I've also removed some duplicate code by moving the initialization step for two of the test methods into a shared method.

PiperOrigin-RevId: 516326606