Bump compose / ktx versions to alpha.4 after the 4.15.1 Glide release.
These need to be bumped separately because we're not currently doing dot releases the way we are with the rest of Glide. The rest of Glide can just use 4.15.1, but we don't want to add the complexity of 1.0.0-alpha.3.1, so we have to bump the version with each dot release.

PiperOrigin-RevId: 516360517