Pass an executor into ChromiumRequestSerializer
ChromiumRequestSerializer creates an single thread executor
for executing callbacks from cronet. In tests, we want to track
all threads related to Glide and cronet through ActiveResource, so
that the test result is deterministic. So we pass an active resource
thread pool executor to ChromiumRequestSerializer, and the active resource
will be tracked by ActiveResourceRule.

PiperOrigin-RevId: 518069577