Set Glide's fragmentx version back to 1.3.6
I'm not sure how I managed to update this to a newer version without breaking the build, but it now appears to be broken. 1.5+ and I think 1.4+ require the usage of java8's desugaring. Glide's v4 doesn't currently do this, so we need to stay on an older version of the Fragment libraries.

Fixes #5086

PiperOrigin-RevId: 519745378