Avoid stripping disable hardware bitmaps on O.
For a while we've stripped some experimental methods when they're exported to GitHub. This prevents them from being included in a public release, which in turn prevents them from causing binary incompatibility issues between different versions of Glide if the methods are misused.

However it also creates a maintenance burden where the internal and external versions of Glide are not the same. It adds some confusion for external developers because necessarily some number of useless classes, imports or variables are left in the exported version.

Rather than rely on explicit stripping, we'll instead document and deprecate the experimental methods and continue to remove them when they're no longer useful. This is more transparent and easier to maintain, at the risk of some additional binary incompatibility issues in the future. Given that it's very unlikely that people will use these methods, especially in libraries, this seems like an ok tradeoff.

PiperOrigin-RevId: 520157449