Open source Glide's FirstFrame waiter
This class helps avoid some issues with hardware Bitmaps on Android O and P. It's been in use in production internally for a while for these older devices. Even if we don't use it on Q+ and we decide there are too many bugs with hardware Bitmaps on O, it's still useful to enable on P devices.

PiperOrigin-RevId: 520157997