Fix errors generating Glide's documentation with dokka
Prior to this change, the build fails because dokka fails to find the
dokka task on a bunch of submodules. That's happening because we
deliberately disabled the dokka task on irrelevant submodules.

I have no idea how this worked originally. I even went back to my
original commit (e5a29beb96d9fc8a31e9f0adc650f6b2bd09c4ec) and tried
building with dokka and it failed. But clearly 4.14's javadocs on the
website were built with dokka.

The fix seems to be to always add the dokka task, but to simply set it
to disabled for submodules that aren't useful for developers. That's a
relatively minor change that I've made here.