Remove GlidePainter in favor of Modifier nodes / Flows
Using a custom Modifier node instead of a Painter follows a
recommendation from the Compose team. It will allow us or
library users to compose Glide's modifier in the future for animations
of other effects.  For now we do not actually expose the Modifier
directly.

This change adds a GlideSubcomposition that uses' Glide's flows
integration to allow subcomposition based on image load state.
Subcomposition allows us to deprecate the expensive Composable
placeholder variants and also allows users to implement complex layouts
or animations.

In a subsequent change we'll explore adding a default crossfade and/or
exposing the Glide modifier node so that users of the library can
compose modifiers themselves.