Use onIdle to avoid a race in FlowTests
Glide's executor thread notifies the flow in a loop while holding a lock
on the resource. When it finishes, it releases the lock.

If the flow wins the race and runs before Glide's executor releases the
lock, the resource will not be recycled in the remainder of the test
method. If the executor wins the race and releases the lock first, then
the resource will be recycled in the rest of the test method.

To fix this race, I've used Espresso's onIdle and idling resources,
similar to the compose rule.