Remove unused <uses-sdk android:minSdkVersion="..." />
Issue id: GradleOverrides
This minSdkVersion value (...) is not used; it is always overridden by the value specified in the Gradle build script (...)
Inspection info: The value of (for example) minSdkVersion is only used if it is not specified in the build.gradle build scripts. When specified in the Gradle build scripts, the manifest value is ignored and can be misleading, so should be removed to avoid ambiguity.