Make Placeholder a type that can be convered to Painter
This removes the deprecated Composable variant of Placeholder. Users who
require that functionality should instead use GlideSubcomposition. It's
useful to remove this now because it allows for a much simpler
Placholder interface and it unblocks us from moving most of GlideImage
into GlideModifier. In turn that will let us expose GlideModifier as a
real composable API component.