avif: Use the correct bucket for GlideModule
Instead of prepending to the default prepend_all bucket, prepend
the Avif integration decoders to the Bitmap bucket. Also add a
configuration for decoding into a BitmapDrawable.

This prevents some incorrect interactions with the Downsampler
when dealing with the animated drawables (if the Avif integration
is in the prepend_all bucket, it prevents GIF animations from
playing back since they are returned as a BitmapDrawable as
explained in issue #5051).

Fixes #5051