Fix `OfResourceId` to allow all Drawable types.
The fine print in `androidx.compose.ui.res.painterResource` says that it
only works for certain kinds of Drawables (Vector and Bitmap), which
isn't reflected in GlideImage.

There's a potential small performance penalty for vector images
(`painterResource` has a hacky fast-path for that), but such is the cost
of working for all Drawables.