Fix deadlock in SingleRequest.clear by using the Engine as the requestLock.
This occurs because CallResourceReady only acquires 2/3 of the necessary locks to perform its runnable (SingleRequest.requestLock & EngineJob.this), and is missing the Engine.this lock.

In this change, I changed SingleRequest.requestLock to always be Engine.this.

Note that change only affects uses of SingleRequest built through RequestBuilder. Users of the public API SingleRequest.obtain are unaffected by this change.

If this fix works as intended, we can consider providing a new SingleRequest.obtain API with the requestLock param removed and deprecating the current SingleRequest.obtain.

PiperOrigin-RevId: 817821083