Limit the size of LruBitmapPool/MemoryCache when app is in background
When an app is moved to background, onTrimMemory() will be called, and
LruBitmapPool/MemoryCache will be trimmed. However, the app could keep
running and bitmaps will be retired to MemoryCache and LruBitmapPool.

This change allows limit of the sizes of LruBitmapPool/MemoryCache (by
MemoryCategory) when the app is in background, and restore the sizes
to normal when it's moved out of background.